{{cookiecutter.project_slug}}/requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
{%- if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %}
rcssmin==1.1.0 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
{%- else %}
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
{%- endif %}
{%- endif %}
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
{%- endif %}
redis==4.5.4  # https://github.com/redis/redis-py
{%- if cookiecutter.use_docker == "y" or cookiecutter.windows == "n" %}
hiredis==2.2.2  # https://github.com/redis/hiredis-py
{%- endif %}
{%- if cookiecutter.use_celery == "y" %}
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
{%- if cookiecutter.use_docker == 'y' %}
flower==1.2.0  # https://github.com/mher/flower
{%- endif %}
{%- endif %}
{%- if cookiecutter.use_async == 'y' %}
uvicorn[standard]==0.21.1  # https://github.com/encode/uvicorn
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django>=4.2.5,<5  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ>=0.10.0,<1  # https://github.com/joke2k/django-environ
django-model-utils>=4.3.1,<5  # https://github.com/jazzband/django-model-utils
django-allauth>=0.54.0,<1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-cookie-consent==0.4.0
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
django-compressor==4.3.1  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-jazzmin>=2.6.0,<3
django-redis>=5.2.0,<6  # https://github.com/jazzband/django-redis

# Django REST Framework
djangorestframework>=3.14.0,<4  # https://github.com/encode/django-rest-framework
django-cors-headers>=4.2.0,<5  # https://github.com/adamchainz/django-cors-headers

{%- if cookiecutter.frontend_pipeline == 'Webpack' %}

django-webpack-loader>=2.0.1,<3 # https://github.com/django-webpack/django-webpack-loader
{%- endif %}

# Wagtail
# ------------------------------------------------------------------------------
wagtail>=5.1.1,<6  # https://github.com/wagtail/wagtail
wagtail-metadata>=5.0.0,<6
